refactor(lib): moved the argument scope to individual module to handle the Parser and Handler classes
