+++ lib/keys.rb
@@ -8,7 +8,7 @@ require_relative './core/utils/swift/package'
 require_relative './core/utils/swift/swift'
 require_relative './core/utils/swift/xcframework'
 require_relative './core/utils/openssl/cipher'
-require_relative './core/console/argument_parser'
+require_relative './core/console/arguments/parser'
 
 module SecureKeys
   class Generator
@@ -20,7 +20,7 @@ module SecureKeys
 
     def initialize
       # Configure the argument parser
-      SecureKeys::Core::Console::ArgumentParser.new
+      SecureKeys::Core::Console::Argument::Parser.new
 
       puts "🔔 You're using a custom delimiter '#{SecureKeys::Globals.key_delimiter}'" unless SecureKeys::Globals.key_delimiter.eql?(SecureKeys::Globals.default_key_delimiter)
       puts "🔔 You're using a custom key access identifier '#{SecureKeys::Globals.key_access_identifier}'" unless SecureKeys::Globals.key_access_identifier.eql?(SecureKeys::Globals.default_key_access_identifier)
